Remove distributivity laws from `GCDMonoid` type class.
These laws are now part of the `DistributiveGCDMonoid` type class.

Prior to this commit, these laws were stated /conditionally/, for types
that are instances of both `GCDMonoid` and `Cancellative`.

However, while cancellativity of `<>` is a /sufficient/ condition for
the `gcd` operation to be distributive (when considering the existing
laws for `gcd`), it's actually not a /necessary/ condition for
distributivity.

There are many non-cancellative GCD monoids that can be shown to
satisify the distributivity laws for `gcd`, such as:

- set monoids with:
    - a `<>`  operation that corresponds to `union`;
    - a `gcd` operation that corresponds to `intersection`.

- natural number monoids with:
    - a `<>`  operation that corresponds to `*`;
    - a `gcd` operation that corresponds to `Prelude.gcd`.