Fix erroneous redundant import warnings in `Data.Monoid.LCM`.
Certain definitions must be imported in order for Haddock to be able to
construct hyperlinks within documentation, but are not actually used in
any function.

Frustratingly, GHC (and other tools such as HLS) mark these imports as
redundant, even though they're actually needed by Haddock.

This commit adds references to just these definitions (and no others)
to prevent them from being marked as redundant imports. This makes it
possible to compile the module with `-Wall -Werror`.