Use `extra-doc-files` for documentation included in cabal file.
More recent cabal formats introduced support for `extra-doc-files` as a
way to specify documentation that is included with a package.

If documentation files are included within an `extra-source-files`
section, `cabal check` will produce the following warnings:

```
Warning: Please consider moving the file 'CHANGELOG.md' from the
'extra-source-files' section of the .cabal file to the section
'extra-doc-files'.
```
```
Warning: Please consider moving the file 'README.md' from the
'extra-source-files' section of the .cabal file to the section
'extra-doc-files'.
```